переписал метод удаления user. Теперь корректно обрбатывает каскадное удаление

+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmDbStorage.java
@@ -141,6 +141,9 @@ public class FilmDbStorage implements FilmStorage {
 
     @Override
     public void delete(Long id) {
+        if (findById(id).isEmpty()) {
+            throw new NotFoundException("Фильм с id " + id + " не найден");
+        }
         deleteGenresByFilmId(id);
         deleteLikesByFilmId(id);
         String query = "delete from films where film_id = ?";
@@ -152,8 +155,8 @@ public class FilmDbStorage implements FilmStorage {
             return new HashMap<Long, Set<Genre>>();
         }
         String placeholder = Stream.generate(() -> "?")
-                                   .limit(filmIds.size())
-                                   .collect(Collectors.joining(", "));
+                .limit(filmIds.size())
+                .collect(Collectors.joining(", "));
         String query = "select g.genre_id, g.name, fg.film_id from genres g join film_genres fg on fg.genre_id = g.genre_id where fg.film_id in (" + placeholder + ")";
         HashMap<Long, Set<Genre>> genresByFilmId = new HashMap<>();
         jdbc.query(query, rs -> {
@@ -162,7 +165,7 @@ public class FilmDbStorage implements FilmStorage {
             genre.setId(rs.getInt("genre_id"));
             genre.setName(rs.getString("name"));
             genresByFilmId.computeIfAbsent(filmId, id -> new HashSet<>())
-                          .add(genre);
+                    .add(genre);
         }, filmIds.toArray());
         return genresByFilmId;
     }
@@ -184,15 +187,15 @@ public class FilmDbStorage implements FilmStorage {
             return new HashMap<Long, Set<Long>>();
         }
         String placeholder = Stream.generate(() -> "?")
-                                   .limit(filmIds.size())
-                                   .collect(Collectors.joining(", "));
+                .limit(filmIds.size())
+                .collect(Collectors.joining(", "));
         String query = "select user_id, film_id from film_likes where film_id in (" + placeholder + ")";
         HashMap<Long, Set<Long>> likesByFilmId = new HashMap<>();
         jdbc.query(query, rs -> {
             Long filmId = rs.getLong("film_id");
             Long userId = rs.getLong("user_id");
             likesByFilmId.computeIfAbsent(filmId, id -> new HashSet<>())
-                         .add(userId);
+                    .add(userId);
         }, filmIds.toArray());
         return likesByFilmId;
     }
@@ -219,12 +222,6 @@ public class FilmDbStorage implements FilmStorage {
     public void removeLike(Long filmId, Long userId) {
         String query = "delete from film_likes where film_id = ? and user_id = ?";
         jdbc.update(query, filmId, userId);
-    }
-
-    @Override
-    public void removeLikesByUser(Long userId) {
-        String query = "delete from film_likes where user_id = ?";
-        jdbc.update(query, userId);
     }
 
     @Override
